[mlir][tensor] Add the tensor.tanspose op
AI frameworks typically have transpose ops on tensors. Previously
we can only use a `linalg.generic` with one input and permuation
indexing map to represent them. That is quite some struct to
represent and pattern match.

So this patch introduces a transpose op to serve a mechanical
importing/lowering for them, and as a tensor counterpart for
`memref.transpose`. This allows us to keep the high level semantics
of transposing until when we need to lower it.